+++ esp-hal/CHANGELOG.md
@@ -21,11 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `Dpi` driver and `DpiTransfer` now have a `Mode` type parameter. The driver's asyncness is determined by the asyncness of the `Lcd` used to create it. (#2526)
 - `dma::{Channel, ChannelRx, ChannelTx}::set_priority` for GDMA devices (#2403)
 - `SystemTimer::set_unit_count` & `SystemTimer::configure_unit` (#2576)
+- `SystemTimer::set_unit_value` & `SystemTimer::configure_unit` (#2576)
 
 ### Changed
 
 - `SystemTimer` no longer uses peripheral ref (#2576)
-- `SystemTimer::now` has been renamed `SystemTimer::unit_count(Unit)` (#2576)
+- `SystemTimer::now` has been renamed `SystemTimer::unit_value(Unit)` (#2576)
 
 ### Fixed